+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- **[BREAKING]** `formService.getFormElementOptionsSetOptions()` parameters
+  ```diff
+  const result = await formService.getFormElementOptionsSetOptions(
+    formElementOptionsSet,
+  - form.formsAppEnvironmentId,
+  + {
+  +   formsAppEnvironmentId: form.formsAppEnvironmentId,
+  +   formId: form.id,
+  + },
+    abortSignal,
+  )
+  ```
+
 ## [21.1.1] - 2025-03-19
 
 ## [21.1.0] - 2025-03-18